Relocated more forecast chart utils
